Modularize models and add timestamps to models
