+++ backend/app/models/user.py
@@ -1,0 +1,24 @@
+"""User model"""
+from flask_bcrypt import bcrypt
+from datetime import datetime
+from app import db
+
+class User(db.Model):
+    """Users model"""
+    __tablename__ = 'users'
+
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(80), unique=True, nullable=False)
+    password_hash = db.Column(db.String(128), nullable=False)
+    created_at = db.Column(db.DateTime, default=datetime.now(), nullable=False)
+    updated_at = db.Column(db.DateTime, default=datetime.now(), nullable=False)
+    deleted_at = db.Column(db.DateTime, nullable=True)
+
+    def __init__(self, username, password):
+        """Initialize user"""
+        self.username = username
+        self.password_hash = bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode('utf-8')
+
+    def check_password(self, password):
+        """Check user password"""
+        return bcrypt.checkpw(password.encode('utf-8'), self.password_hash.encode('utf-8'))
